rename all folders and scripts so that they no longer start with a number
