+++ .editorconfig
@@ -1,18 +1,12 @@
+# EditorConfig is awesome: https://EditorConfig.org
+
+# top-most EditorConfig file
 root = true
 
 [*]
 indent_style = space
-indent_size = 2
-
+indent_size = 4
 end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
-
-[*.{py,cfg,toml,rst}]
-indent_style = space
-indent_size = 4
-
-[Makefile]
-indent_style = tab
-indent_size = 4
