+++ python/pyproject.toml
@@ -5,18 +5,18 @@ description = "Open Discourse"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
-    "beautifulsoup4>=4.12.3",
+    "beautifulsoup4>=4.13.3",
     "dicttoxml>=1.7.16",
-    "levenshtein>=0.25.1",
+    "levenshtein>=0.27.1",
     "numpy>=2.0.0",
     "pandas>=2.2.2",
-    "psycopg2>=2.9.9",
     "regex>=2024.5.15",
     "requests>=2.32.3",
-    "sqlalchemy>=2.0.31",
+    "sqlalchemy>=2.0.39",
     "tqdm>=4.62.3",
-    "pyparsing>=3.2.0",
-    "pydantic>=2.9.2",
+    "pyparsing>=3.2.1",
+    "pydantic>=2.10.6",
+    "psycopg2-binary>=2.9.10",
 ]
