+++ python/src/open_discourse/topic_modelling/upload_data_cubes.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import psycopg2
 
-import open_discourse.definitions.path_definitions as path_definitions
+import open_discourse.definitions.path as path
 
 
 def upload_dims(cur, dims, shape, schema):
@@ -112,9 +112,9 @@ conn = psycopg2.connect(
 cur = conn.cursor()
 
 
-data_cube = pd.read_pickle(path_definitions.FINAL / "data_cube.pkl")
-weight_cube = pd.read_pickle(path_definitions.FINAL / "weight_cube.pkl")
-dims = pd.read_pickle(path_definitions.FINAL / "dims.pkl")
+data_cube = pd.read_pickle(path.FINAL / "data_cube.pkl")
+weight_cube = pd.read_pickle(path.FINAL / "weight_cube.pkl")
+dims = pd.read_pickle(path.FINAL / "dims.pkl")
 shape = data_cube.shape
 schema = "lda_group"
 
@@ -136,9 +136,9 @@ recursive_upload(
 conn.commit()
 
 
-data_cube = pd.read_pickle(path_definitions.FINAL / "politician_data_cube.pkl")
-weight_cube = pd.read_pickle(path_definitions.FINAL / "politician_weight_cube.pkl")
-dims = pd.read_pickle(path_definitions.FINAL / "politician_dims.pkl")
+data_cube = pd.read_pickle(path.FINAL / "politician_data_cube.pkl")
+weight_cube = pd.read_pickle(path.FINAL / "politician_weight_cube.pkl")
+dims = pd.read_pickle(path.FINAL / "politician_dims.pkl")
 shape = data_cube.shape
 schema = "lda_person"
